+++ src/client/sprites/Button.ts
@@ -1,0 +1,156 @@
+import { canvas, ctx, penCtx } from '../canvas.ts';
+import Sprite, { type BoundingBox, type SpriteOptions } from '../Sprite.ts';
+
+export interface ButtonOptions extends SpriteOptions {
+    content?: string;
+    fontColor?: string;
+    fontFamily?: string;
+    fontSize?: number;
+
+    width?: number;
+    height?: number;
+    backgroundColor?: string;
+
+    outlineColor?: string;
+    outlineWidth?: number;
+}
+
+export default class Button extends Sprite {
+
+    public discriminant = 'button';
+
+    public content: string;
+    public fontColor: string;
+    public fontFamily: string;
+    public fontSize: number;
+
+    private font: string;
+
+    public width: number;
+    public height: number;
+    public backgroundColor: string;
+
+    public outlineColor: string;
+    public outlineWidth: number;
+
+    public getBoundingBox(): BoundingBox {
+
+        ctx.save();
+        
+        ctx.font = this.font;
+
+        const metrics = ctx.measureText(this.content);
+        const width = Math.max(metrics.width, this.width);
+        const height = Math.max(metrics.actualBoundingBoxAscent + metrics.actualBoundingBoxDescent, this.width);
+
+        ctx.restore();
+
+        return {
+            x: this.x,
+            y: this.y,
+            width: Math.hypot(width, height),
+            height: Math.hypot(width, height)
+        };
+    }
+
+    public getPath(): Path2D {
+        const path = new Path2D();
+
+        path.rect(
+            -this.width / 2,
+            -this.height / 2,
+            this.width,
+            this.height
+        );
+
+        return path
+    }
+
+    public draw(stamping?: boolean) {
+        const c = stamping ? penCtx : ctx;
+
+        c.save();
+
+        const cX = this.x + canvas.width / 2;
+        const cY = -this.y + canvas.height / 2;
+        
+        c.translate(cX, cY);
+        c.rotate(this.toRadians(this.dir));
+
+        c.fillStyle = this.backgroundColor;
+        c.strokeStyle = this.outlineColor;
+        c.lineWidth = this.outlineWidth;
+
+        const path = this.getCachedPath();
+        
+        c.fill(path);
+        if (this.outlineWidth) c.stroke(path);
+
+        c.font = this.font;
+        c.fillStyle = this.fontColor;
+        c.textAlign = 'center';
+        c.textBaseline = 'middle';
+
+        c.fillText(this.content, 0, 0);
+
+        c.restore();
+    }
+
+    public create(options?: ButtonOptions): this {
+        return new Button(options) as this;
+    }
+
+    // Methods
+
+    public setContent(content: string) {
+        this.content = content;
+        this.invalidatePath();
+        this.refresh();
+    }
+
+    public setBackgroundColor(color: string) {
+        this.backgroundColor = color;
+        this.refresh();
+    }
+
+    public setFontColor(color: string) {
+        this.fontColor = color;
+        this.refresh();
+    }
+
+    public setFontSize(fontSize: number) {
+        this.fontSize = fontSize;
+        this.font = `${this.fontSize}px ${this.fontFamily}`;
+        this.invalidatePath();
+        this.refresh();
+    }
+
+    public setFontFamily(fontFamily: string) {
+        this.fontFamily = fontFamily;
+        this.font = `${this.fontSize}px ${this.fontFamily}`;
+        this.invalidatePath();
+        this.refresh();
+    }
+
+    // Constructor
+
+    constructor(options?: ButtonOptions) {
+        super(options);
+
+        this.content = options?.content ?? '';
+        this.fontColor = options?.fontColor ?? 'black';
+        this.fontFamily = options?.fontFamily ?? 'Arial';
+        this.fontSize = options?.fontSize ?? 16;
+
+        this.font = `${this.fontSize}px ${this.fontFamily}`;
+
+        this.width = options?.width ?? this.content.length * this.fontSize + 10;
+        this.height = options?.height ?? this.fontSize + 10;
+        this.backgroundColor = options?.backgroundColor ?? 'rgb(204, 204, 204)';
+
+        this.outlineColor = options?.outlineColor ?? 'black';
+        this.outlineWidth = options?.outlineWidth ?? 0;
+
+        this.draw();
+    }
+}