+++ src/client/types/Matricies.ts
@@ -1,0 +1,17 @@
+export type Mat2<T = number> = [
+    [T, T],
+    [T, T]
+];
+
+export type Mat3<T = number> = [
+    [T, T, T],
+    [T, T, T],
+    [T, T, T]
+];
+
+export type Mat4<T = number> = [
+    [T, T, T, T],
+    [T, T, T, T],
+    [T, T, T, T],
+    [T, T, T, T]
+];