+++ src/client/types/Vectors.ts
@@ -1,3 +1,3 @@
-export type Vec2 = [number, number];
-export type Vec3 = [number, number, number];
-export type Vec4 = [number, number, number, number];
+export type Vec2<T = number> = [T, T];
+export type Vec3<T = number> = [T, T, T];
+export type Vec4<T = number> = [T, T, T, T];